requirements.txt
django==1.7.3
djangorestframework==3.0.3
django-rest-swagger==0.2.8